--- source/databricks/calculation_engine/tests/business_logic_tests/features/energy_calculations/flex_consumption_per_ga_and_brp_and_es/readme.md
@@ -1,1 +1,0 @@
-# flex_consumption_per_ga_and_brp_and_es
