+++ source/databricks/calculation_engine/tests/calculation_logic_tests/features/energy_calculations/flex_consumption_per_ga_and_es/readme.md
@@ -1,0 +1,15 @@
+# Calculate flex consumption per grid area and energy supplier
+
+The purpose of this test is to verify that the flex consumption is calculated correctly for
+grid area 804.
+
+Period:
+
+            Period_Start                         Period_End
+             2023-01-31                          2023-02-01
+              23:00:00                            23:00:00
+              ---|-----------------------------------|---
+        
+    MPP1 (E17)   |----------------------------------------------
+    MPP2 (E18)   |----------------------------------------------
+    MPP3 (E17) -------------------------------------------------
