+++ source/databricks/calculation_engine/tests/calculation_logic_tests/features/energy_calculations/flex_consumption_per_ga_and_es/states/scenario_state.py
@@ -12,11 +12,10 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from pyspark.sql import DataFrame, SparkSession
-
-from business_logic_tests.features.energy_calculations.results_dataframe import (
+from calculation_logic_tests.features.energy_calculations.results_dataframe import (
     create_result_dataframe,
 )
+from pyspark.sql import DataFrame, SparkSession
 
 
 def get_expected(*args) -> DataFrame:  # type: ignore
