+++ source/databricks/calculation_engine/tests/calculation_logic_tests/scenario_factory.py
@@ -17,10 +17,10 @@ import os
 from typing import Callable
 from unittest.mock import Mock
 
+from calculation_logic_tests.correlations import get_correlations
+from calculation_logic_tests.create_calculation_args import create_calculation_args
 from pyspark.sql import SparkSession, DataFrame
 
-from business_logic_tests.correlations import get_correlations
-from business_logic_tests.create_calculation_args import create_calculation_args
 from package.calculation import PreparedDataReader
 from package.calculation.CalculationResults import (
     CalculationResultsContainer,
